requirements.txt
flask
protobuf
pytest
pyyaml
deepdiff

# OpenTelemetry Tests
google==3.0.0
opentelemetry-proto
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
grpcio
